Cleaned up icons in source/target columns
