+++ ui/packages/atlasmap/src/Atlasmap/AtlasmapProvider.tsx
@@ -1,3 +1,11 @@
+import {
+  DocumentInitializationModel,
+  DocumentType,
+  InspectionType,
+  MappingSerializer,
+  MappingUtil,
+  search,
+} from "@atlasmap/core";
 import React, {
   createContext,
   FunctionComponent,
@@ -7,33 +15,38 @@ import React, {
   useReducer,
 } from "react";
 import { debounceTime } from "rxjs/operators";
-
-import {
-  MappingUtil,
-  search,
-  MappingSerializer,
-  InspectionType,
-  DocumentType,
-  DocumentInitializationModel,
-} from "@atlasmap/core";
-
 import { IAtlasmapDocument, IAtlasmapField } from "../Views";
+import {
+  DataActionPayload,
+  dataReducer,
+  IDataState,
+  initDataState,
+  initNotificationsState,
+  INotificationsState,
+  notificationsReducer,
+} from "./reducers";
 import {
   addToCurrentMapping,
   createConstant,
   createMapping,
+  createNamespace,
   createProperty,
   deleteAtlasFile,
   deleteConstant,
+  deleteNamespace,
   deleteProperty,
   deselectMapping,
   documentExists,
   editConstant,
+  editNamespace,
   editProperty,
   enableCustomClass,
+  errorInfoToNotification,
   executeFieldSearch,
   exportAtlasFile,
   fromDocumentDefinitionToFieldGroup,
+  fromFieldToIFieldsNode,
+  fromMappedFieldToIMappingField,
   fromMappingDefinitionToIMappings,
   fromMappingModelToImapping,
   getMappingActions,
@@ -42,12 +55,12 @@ import {
   getMultiplicityActions,
   handleActionChange,
   handleIndexChange,
-  handleNewTransformation,
-  handleTransformationChange,
-  handleTransformationArgumentChange,
-  handleRemoveTransformation,
-  handleMultiplicityChange,
   handleMultiplicityArgumentChange,
+  handleMultiplicityChange,
+  handleNewTransformation,
+  handleRemoveTransformation,
+  handleTransformationArgumentChange,
+  handleTransformationChange,
   importAtlasFile,
   initializationService,
   mappingExpressionAddField,
@@ -58,32 +71,17 @@ import {
   mappingExpressionRemoveField,
   newMapping,
   onFieldPreviewChange,
-  toggleExpressionMode,
+  removeFromCurrentMapping,
+  removeMappedFieldFromCurrentMapping,
   removeMapping,
   resetAtlasmap,
   selectMapping,
+  toggleExpressionMode,
   toggleMappingPreview,
   toggleShowMappedFields,
   toggleShowUnmappedFields,
   trailerId,
-  removeFromCurrentMapping,
-  removeMappedFieldFromCurrentMapping,
-  fromMappedFieldToIMappingField,
-  errorInfoToNotification,
-  createNamespace,
-  editNamespace,
-  deleteNamespace,
-  fromFieldToIFieldsNode,
 } from "./utils";
-import {
-  INotificationsState,
-  IDataState,
-  initDataState,
-  initNotificationsState,
-  dataReducer,
-  notificationsReducer,
-  DataActionPayload,
-} from "./reducers";
 
 // the document payload with get from Syndesis
 export interface IExternalDocumentProps {
@@ -278,15 +276,15 @@ export const AtlasmapProvider: FunctionComponent<IAtlasmapProviderProps> = ({
   );
 
   const onSubUpdate = useCallback(
-    function onSubUpdateCb(_caller: string) {
-      // console.log(
-      //   "onUpdates",
-      //   caller,
-      //   "initialized",
-      //   initializationService.cfg.initCfg.initialized,
-      //   "errors",
-      //   initializationService.cfg.initCfg.initializationErrorOccurred,
-      // );
+    function onSubUpdateCb(caller: string) {
+      console.log(
+        "onUpdates",
+        caller,
+        "initialized",
+        initializationService.cfg.initCfg.initialized,
+        "errors",
+        initializationService.cfg.initCfg.initializationErrorOccurred,
+      );
       onUpdates({
         pending: !initializationService.cfg.initCfg.initialized,
         error: initializationService.cfg.initCfg.initializationErrorOccurred,
