--- ui/packages/atlasmap/src/Atlasmap/dialogs/useExportCatalogDialog.story.tsx
@@ -1,24 +1,0 @@
-import React, { useCallback, ReactElement } from "react";
-
-import { useToggle, ExportCatalogDialog } from "../../UI";
-import { useAtlasmap } from "../AtlasmapProvider.story";
-
-export function useExportCatalogDialog(): [ReactElement, () => void] {
-  const { state, toggleOn, toggleOff } = useToggle(false);
-  const { exportAtlasFile } = useAtlasmap();
-  const onExportAtlasFile = useCallback(
-    (filename: string) => {
-      exportAtlasFile(filename);
-      toggleOff();
-    },
-    [exportAtlasFile, toggleOff],
-  );
-  const dialog = (
-    <ExportCatalogDialog
-      isOpen={state}
-      onCancel={toggleOff}
-      onConfirm={onExportAtlasFile}
-    />
-  );
-  return [dialog, toggleOn];
-}
