+++ ui/packages/atlasmap/src/Atlasmap/utils/toolbar.ts
@@ -6,7 +6,6 @@ import {
   ErrorType,
   InspectionType,
 } from "@atlasmap/core";
-
 import { getDocDef, importInstanceSchema, removeDocumentRef } from "./document";
 
 /**
