+++ ui/packages/atlasmap/src/Views/index.ts
@@ -2,7 +2,8 @@ export * from "./ColumnMapperView";
 export * from "./FreeViewMapper";
 export * from "./MappingDetailsView";
 export * from "./MappingTableView";
-export * from "./NamespaceTableView";
 export * from "./models";
+export * from "./NamespaceTableView";
 export * from "./SourceMappingTargetView";
+export * from "./SourceMappingTargetXformView";
 export * from "./SourceTargetView";
